renamed diagram_gen to plotting
